move generate route to source_app/views.py
