+++ securedrop/source_templates/index.html
@@ -59,7 +59,7 @@
             </div>
 
             <div class="index-column index-left index-bottom">
-              <a href="/generate" id="submit-documents-button" class="sd-button btn alt">
+              <a href="{{ url_for('main.generate') }}" id="submit-documents-button" class="sd-button btn alt">
                 <img class="icon off-hover" src="{{ url_for('static', filename='i/font-awesome/cloud-upload-white.png') }}" width="20px" height="14px">
                 <img class="icon on-hover" src="{{ url_for('static', filename='i/font-awesome/cloud-upload-blue.png') }}" width="20px" height="14px">
                 {{ gettext('SUBMIT DOCUMENTS') }}
